test: ConfirmPasswordTest 도메인 로직, UI 로직 테스트 분리
