+++ app/src/androidTest/java/nextstep/signup/ConfirmPasswordTest.kt
@@ -3,20 +3,20 @@ package nextstep.signup
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.test.junit4.createComposeRule
+import androidx.compose.ui.test.onNodeWithText
 import nextstep.signup.ui.component.SignUpTextFieldComponent
 import nextstep.signup.ui.model.ConfirmPassword
 import nextstep.signup.ui.model.Password
-import nextstep.signup.ui.model.SignUpState
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 
-class ConfirmPasswordTest {
+class ConfirmPasswordUITest {
 
     @get:Rule
     val composeTestRule = createComposeRule()
 
-    private val password = Password(text = "Password1")
+    private val password = Password(text = "password")
     private val confirmPassword = mutableStateOf("")
 
     @Before
@@ -32,32 +32,27 @@ class ConfirmPasswordTest {
     }
 
     @Test
-    fun 확인비밀번호가_빈_값이면_Blank_상태를_반환한다() {
-        // when
+    fun 확인비밀번호가_빈_값일_때_Blank_상태를_반환한다() {
         confirmPassword.value = ""
-
-        // then
-        val result = ConfirmPassword(password, confirmPassword.value).validState()
-        assert(result is SignUpState.Blank)
+        composeTestRule
+            .onNodeWithText(stringResource(R.string.password_confirm_label))
+            .assertExists()
     }
 
     @Test
-    fun 확인비밀번호가_원래비밀번호와_다르면_Confirm_에러를_반환한다() {
-        // when
-        confirmPassword.value = "WrongPassword"
-
-        // then
-        val result = ConfirmPassword(password, confirmPassword.value).validState()
-        assert(result is SignUpState.InValid.Confirm)
+    fun 확인비밀번호가_원래비밀번호와_다를_때_Confirm_에러를_반환한다() {
+        confirmPassword.value = "wrongPassword"
+        composeTestRule
+            .onNodeWithText(stringResource(R.string.password_confirm_label))
+            .assertExists()
     }
 
     @Test
-    fun 확인비밀번호가_유효할_경우_유효_상태인_Valid를_반환한다() {
-        // when
-        confirmPassword.value = "Password1"
-
-        // then
-        val result = ConfirmPassword(password, confirmPassword.value).validState()
-        assert(result is SignUpState.Valid)
+    fun 확인비밀번호가_유효할_때_유효_상태인_Valid를_반환한다() {
+        confirmPassword.value = "password"
+        composeTestRule
+            .onNodeWithText(stringResource(R.string.password_confirm_label))
+            .assertExists()
     }
 }
+
